Allow rpm.t to compile EPEL-7 for master and release-3.5 branches
Unsure if release-3.4 branch can handle EPEL-7 yet.  Will update
this when I find out.

+++ regression_test.sh
@@ -67,9 +67,11 @@ GERRIT_CR=`metadata_retriever.py change_req 2>/dev/null`
 # Extract Git ref if there is one
 GIT_REF=`metadata_retriever.py change_ref 2>/dev/null`
 
-# TODO: Remove this once mock works with EPEL-7 again
-# Workaround for some GlusterFS branches not compiling EPEL-7 yet
-rm -f /etc/mock/epel-7-x86_64.cfg
+# TODO: Remove this once release-3.4 branch compiles EPEL-7 ok
+# Workaround for GlusterFS release-3.4 branch not compiling EPEL-7 yet
+if [ x"$GLUSTERFS_BRANCH" = x'release-3.4' ]; then
+    rm -f /etc/mock/epel-7-x86_64.cfg
+fi
 
 # Prepare for building Gluster and running the tests
 mkdir -p /d/archived_builds >> ${COMMAND_LOG} 2>&1
